Add support for language servers to request configuration state (#1608)
This change builds on #1606 to support the language server's configuration loop. These changes are based on discussions I had with Kamiel, around how the language server will handle configuration messages.

The language server does not expect any payload with `workspace/didChangeConfiguration` messages. When received, the language server will immediately send the client a `workspace/configuration` request. In this change, the Toolkit no longer passes any data through didChangeConfiguration, just an empty object.

The language server expects that `workspace/configuration` messages sent to the Toolkit will be responded with a generic map of string keys to object values. This change sets up the first configuration that will be handled by the language server, which indicates if suggestions should filter out those containing license attribution. The Toolkit has been updated to return a response in the form: `{"includeSuggestionsWithCodeReferences":true}`

This is the first handler set up in the Toolkit to respond to a message coming from the language server. I have set up `LspMessageHandler` in a way that can be extended to support additional messages in the future.