Update ToolkitTimer to encapsulate System.Timers.Timer instead of extend it (#1617)
This change updates IToolkitTimer and the ToolkitTimer implementation to be a freestanding Timer implementation that can be stubbed in tests without a reliance on System.Timers.Timer.

The existing implementation has not been used yet, but will be in an upcoming change. The issue that comes with exposure to System.Timers.Timer is that the ElapsedEventArgs class only has an internal constructor. So with a little encapsulation, we still have a drop in replacement to Timer that gives us better code testing flexibility.