Establish a base test for the Connection component (#1618)
The connection component doesn't have any tests yet. Before adding more code to that component, I want to set up the class so that it can be brought under test.

To isolate the GUI related code, I took the approach of moving it to virtual functions that can be stubbed from a test implementation. We can change approaches (like using DI instead) if we like. For now, the objective is to enable ourselves to bring other code in the component under test.

Two thin tests exercising "Sign in" have been added, for the purpose of ensuring the component can be exercised under test. There is a backlog task to fill in more sign in tests.

This change backfills some of the missing Fakes now that we have tests that leverage them.