Auto-refresh tokens for CodeWhisperer (#1620)
This change sets up the CodeWhisperer Connection component to continuously refresh tokens after Sign-In. This way, the user can continuously use CodeWhisperer features without having to re sign-in every hour.

Refreshes only succeed if they do not require user interaction (like performing the SSO login flow). If the system is unable to refresh the token, the CodeWhisperer connection is moved to the expired state, and then the user is signed out. In a future change, the expired state will cause the Toolkit to display a banner, where the user is informed that their connection has expired.