Encapsulate sso token retrieval, increase token refresh test scenarios (#1625)
This change is a follow-up to #1620. It adds testing coverage around the actual refresh handling logic. To do this, the portion of the connection component that obtains SSO Tokens has been encapsulated, so that it can be stubbed out. Without this encapsulation, the tests have a reliance on the local system's sso token cache, which can cause different results between developer and CI systems.