Add opt out setting to share content with CW (#1635)
Builds on top of PR #1608 to add setting to allow users to opt out of having CodeWhisperer train with their code. This setting is sent on Lsp Client initialization and when settings in the option dialog are saved/updated.
The language server expects the configuration state to contain `shareCodeWhispererContentWithAWS` flag to determine the value of the opt out setting to be sent to the CodeWhisperer service.