Add LSP Client State to know when the server is running (#1647)
This change adds plumbing so that other components can determine if the language server is running. In follow-up changes, this plumbing will be consumed to surface the error status in the margin, and to fix some exceptions that occur when trying to access features while the server is not running.