Don't perform operations when the language server is not running (#1652)
If the language server is not running (or initialized) when certain operations are performed, users will see unexpected error messages.

This change updates margin commands so that they will not execute if the language server is not running. It also prevents the Toolkit from pushing configuration changes if the server is not running.