Send telemetry event from language server to backend (#1661)
This change follows a similar pattern as PR #1606 and #1608 to handle telemetry/event notifications sent from the language server. It follows the data contract established here with the language server where a MetricEvent param is sent with the notification.
The toolkit parses the metric event to populate the metric datum which is then emitted to the telemetry backend.

The way it works is to make the base language client raise an event when the notification is received. We then define a separate component (TelemetryEventPublisher) that watches for this event, and emits it after processing(different lsp clients can override the pre-processing behavior as necessary) to the telemetry backend.