Support sending credential start url to server (#1666)
This change follows a similar pattern as PR #1608 and #1661 to handle `getConnectionMetadata` requests sent from the language server. It follows the data contract established here with the language server where a `ConnectionMetadata` object is passed as response back to the server.
The toolkit retrieves the currently signed in connection's start url and passes it back to the server.

The way it works is to make the base language client raise an event when the request is received. We then delegate the handling to `Connection` that watches for this request and populates the current signed in sso start url to the metadata object which is then sent as response to the server.

`ConnectionProperties` has been introduced as a clean way to define and pass auth related properties/data throughout the Connection class