Display notice requiring VS 2022 17.7 or newer in an upcoming release (#1664)
This change informs users running VS 2022 version 17.0-17.6 that the next Toolkit release will require a minimum version of 17.7. The InfoBar has a "learn more" gesture, which links to https://github.com/aws/aws-toolkit-visual-studio/issues/375, and a "don't show again" gesture, which prevents the notice from being seen again.

Much of this was modelled after SupportedVersionStrategy, where we deprecated a whole Major Version of Visual Studio (VS 2017). This is the first time we needed functionality relating to a version within a major version band, so I modeled code on how that one is written, and made it reusable for future usage:

- SunsetNotificationStrategy is a base implementation, which future notices derive from and define logic around their supported versions
- SunsetNotificationSettings backs future sunset notifications, allowing users to suppress seeing the notice again in the future
- SunsetNotificationInfoBar is a generic "Sunset" notification, which is driven by SunsetNotificationStrategy implementations
- SunsetNotificationBarManager is a generic "Sunset" notification orchestrator, which re-drives showing the notice when the VS window is not visible.