Invalidate SSO token if sign in gets stuck in infinite failures (#1685)
PR #1672 updated the AWSSDK package referenced by the toolkit.
Issue: When the AWS SDK updated SSO Token handling, they introduced change that raises an exception if the cached token's client registration time field is null. Previous versions of the AWS SDK would write null to this, resulting in a somewhat backwards incompatible situation (https://github.com/aws/aws-sdk-net/pull/3083/files#r1389714680).

In this scenario, a user would fail to sign-in, and any repeated attempt to sign in would have the same effect. The only resolution would be to delete the cached sso token. Users would not know this, and would be confused why they cannot log in.

Fix: This change improves the Toolkit's sign-in handling to insulate from unexpected SSO credential resolution failures from the AWS SDK. If signing in fails to resolve an SSO Token, the cached token is automatically invalidated (deleted), and the user is told to try again. A subsequent attempt to sign-in is now successful.