Update tests using the UI thread to use TestProjectMockCollection (#1694)
## Problem

Code that works with the UI Thread (usually involving ThreadHelper or JoinableTaskFactory) require portions of the VS SDK to work properly. When this code is put under test, it will not run, because VS SDK bits aren't available. #7 introduced a workaround, but is not reliable. We are about to upgrade the VS SDK, and this could introduce additional stability problems.

## Solution

Switch the affected tests to use a more reliable testing approach introduced in #1587. The majority of the code changes are to set up the JoinableTaskFactory, and ensure test code of interest is running from the UI thread.

Unfortunately, this means we lose test coverage in VS 2019, as this newer approach only supports VS 2022. We have decided to accept this risk, because we believe we will soon end support for VS 2019.