Introduce logic to send inline completion session result to server (#1695)
This is first in a 2 part change to send inline completion session result to server. This change introduces the concept of a suggestion session where the server sends back a session-id to uniquely identify a session along with a list of suggestions (each identified with their own unique id).
In order to communicate data about some user interactions(required for instrumenting telemetry), the `logInlineCompletionSessionResults` notification is sent to the server when a suggestion session completes(i.e when it is accepted or dismissed - will be wired up in a follow-up PR).

This PR instruments the e2e loop of reading the unique id's sent from server, preserving it and stubbing out sending the completion session result back to the server.