Make AwsConnectionManagerTests faster and fix testing bugs (#1718)
The AwsConnectionManagerTests tests take a disproportionately large amount of the overall time running the test suite. This change improves the speed of these tests, which improves our inner development loop, and overall CI build times.

A lot of these tests involve changing AwsConnectionManager state, then waiting for the AwsConnectionManager `ConnectionState` property to change to a certain value. It turns out that there was a flaw in how this was handled -- it did not verify that the state actually changed. Tests [waited for a synchronization signal to fire, but did not verify that the signal fired after reaching a time limit](https://github.com/aws/aws-toolkit-visual-studio-staging/blob/8df3a00/tests/AWSToolkit.Tests/Credentials/Core/AwsConnectionManagerTest.cs#L470-L473):

```
        private void WaitUntilConnectionStateIsStable()
        {
            _connectionStateIsTerminalEvent.WaitOne(10000);
        }
```

The result of `WaitOne` should have been tested for `true`. Many of these tests were simply waiting out the ten seconds, and proceeding. This introduced much delay, as well as testing logic flaws: the tests were passing, but the system under test was not in the intended state, for many of the tests.

An event wrapper task was more recently introduced (#1493), which makes it easer to set these tests up to wait for the intended state. As a result, we're now able to wait for the expected state change, and use a time-based CancellationToken to constrain these tests to an upper limit.

In local development, this change reduced the time taken to run the AwsConnectionManagerTests tests from 80 seconds to 30.

Additional tests were updated, where the testing logic was incorrect.