Process divergence in proposal irrespective of reasons to keep suggestions shown to user updated (#1731)
This change addresses a bug where if user continues typing and the typed prefixes matches a few of the original suggestions(1 out of 3), the UI continues to show 1/3 suggestions available when it should have showed 1/1. This happens because the OnProposalUpdated call is invoked with ReasonForUpdate.UpdatedTypeChar which is not in our currently list of reasons on which we proceed to perform filtering on(as a result even with updated typed prefix, list of suggestions displayed remains unchanged)

This check for reasons was done originally as part of performance improvements as removing that implies filtering is performed on every keystroke. On manual testing after removing the check for list of reasons, no performance impact in terms of typing and viewing suggestions was seen.