Center modal dialogs by using the same ShowModal call as the VS SDK DialogWindow (#1758)
Many modal dialogs presented in the Toolkit often display at a random position, rather than centered with the main Visual Studio window. This is caused when windows aren't properly assigned a Parent or Owner. Assigning WPF dialogs a parent of "the main Visual Studio window" is not trivial. The guidance is to use the [DialogWindow class](https://learn.microsoft.com/en-us/dotnet/api/microsoft.visualstudio.platformui.dialogwindow?view=visualstudiosdk-2022) from the VS SDK, however the Toolkit has many standalone dialogs, implemented in projects that do not have access to the VS SDK.

This change updates the Toolkit ShellProvider's ShowModal method to make the same call that DialogWindow does in order to display modal dialogs (WindowHelper.ShowModal(window)).

The former implementation was from 7+ years ago, and appears to partially do what WindowHelper.ShowModal is doing, so this change should help. I can't guarantee that this will solve the problem 100% of the time, but it should reduce the likelihood that windows are shown off-center.