Remove ToolkitFactory.Instance.Navigator coupling from Lambda publishing (#1756)
One of the reasons the Toolkit's initialization is slow is because we instantiate and populate the AWS Explorer (NavigatorControl) during the extension initialization. The rest of the extension has a lot of tight coupling against NavigatorControl (through the ToolkitFactory.Instance singleton) as it was used as the global representation of accounts and resources.

To fix the slow initialization problem, we need to move NavigatorControl instantiation away from the Toolkit initialization, however we have to first remove all of the existing coupling in order to avoid null access exceptions. This will have to be performed over many isolated changes.

This change removes the Lambda publishing dependency to NavigatorControl. It is able to use the ToolkitContext ConnectionManager instead (which also backs the data used in NavigatorControl).

This change also introduces `ExecuteOnBackgroundThread`, which is a parallel to `ExecuteOnUIThread`. This allows us to run code off the UI thread from modules that don't have access to the VS SDK/JoinableTaskFactory.