Remove unused Build Tasks and Targets relating to old hostedfiles publishing (#1764)
This change removes unused BuildTasks, mostly related to how we used to update/publish hostedfiles contents. Since we've moved to a new publishing pipeline for that content, we no longer use a fair amount of these build tasks. Removing this code helps to make our build definitions more manageable, as there is less code to trip over and try to understand.

For all of the removed code, I have checked that this repo, and the internal code repo do not contain references to it.

Removed BuildTasks:
- DownloadFromS3Task
- UploadFileToS3Task
- UploadReleasesToS3Task
- CompareToS3Task
- UploadHostedFileTask
- UploadHostedFilesTask
- CFInvalidateTask
- EC2QuickLaunchGeneratorTask

Removed build targets:
- compare-hostedfiles *
- upload-hostedfiles *
- upload-hostedfile-ec2-quicklaunch *