Fix S3 bucket location validation for the eu-west-1 region (#1774)
## Problem 

This problem is specific to the "eu-west-1" (Ireland) region. 

To validate the location of the S3 bucket, we use the sdk's `GetBucketLocation` which returns a `LocationConstraint` value rather than a region. The `LocationConstraint` value maps directly to the region value in almost all cases except for the following:
- `us-east-1` where the `LocationConstraint` is returned as null (we already handle this case)
- `eu-west-1` where the `LocationConstraint` can be returned as either `EU` or `eu-west-1` depending on how the bucket was created (this PR handles this case)
  - With both `LocationConstraint` values of `EU` and `eu-west-1`, the bucket actually lives in the region `eu-west-1`
  
## Solution

We need special handling here to account for this "EU" location constraint situation during bucket location validation.