Add GitHub Action and scripting to create release candidate branches (#1778)
This change provides us with a gesture that sets up a release candidate. The gesture is a GitHub action that can be manually (or programmatically) invoked.

The action does the following:
- creates a new release candidate branch for the version of interest
- compiles the release notes for this release candidate
- adjusts the release manifest file consumed by the release pipeline
- pushes the new release candidate branch to the repo

To support this change, the changelog system was revised so that it could emit a custom date instead of today's date. This allows us to create release candidates that we intend to release on a future date.