Adds support for deploying .NET 8 lambdas (#1793)
* Define .NET 8 managed runtime option for Lambda deploy dialog (#1763)

This change adds the .NET 8 managed runtime option into the Lambda publish wizard similar to PR #914

Testing
Verified preview ECR .NET 8 image and zip uploads (with preview access) can be deployed and invoked remotely successfully

* Update Amazon.Lambda.Tools to support .NET 8 container build (#1786)

Pulling in PR aws/aws-extensions-for-dotnet-cli#298 that adds support for using container builds with .NET 8. This PR also confirmed deployments to the upcoming .NET 8 runtime are successful. The testing confirmed function and serverless deployments.

The was also a community PR, aws/aws-extensions-for-dotnet-cli#287, that is being pulled into VS that allows users to configure the interval for polling CloudFormation for status. VS users will be able to take advantage of this feature via the aws-lambda-tools-defaults.json file. I don't see a need for a UI element for configuring this feature. More useful in CI systems.

* Add v11 folder for .NET 8 2022 Lambda blueprints (#1792)

Updates Lambda blueprints for VS2022 to be fetched from msbuild-v11 hosted files location(managed by a separate pipeline).
This change is being made in preparation for a future change that would introduce .NET 8 blueprints. Customers who have not upgraded to the latest version of the toolkit will not have access to the above mentioned blueprints.
VS 2019 or lower versions of VS would continue to target the v9 folder.

---------

Co-authored-by: Norm Johanson <normj@amazon.com>