Update AWS SDK and fix compilation failures (#1800)
This change updates the AWS SDK packages to a recent version. Additional code adjustments have been made to account for some backwards incompatible changes made in the SDK.

One of the main reasons we're updating the AWS SDK is to benefit from the CancellationToken support added to SSO/Token requests. This way the Toolkit won't have threads continuing to poll for SSO tokens for up to ten minutes after the user presses cancel on the SSO Login Flow dialog.