Eliminate UI Thread Lock Risk when writing to output pane (#1801)
David Kean from Microsoft reached out to inform us that some Toolkit users encounter a UI deadlock in their IDE during startup. In the instance reported, it was determined that calls to write text to the output could cause this.

To fix the problem, all code flows to output text have been routed through an async call, which can be fired-and-forgotten, instead of awaiting its completion. This is valid because most of our code doesn't need to wait/block on sending text to the output pane.

The specific deadlock that was brought to our attention is when the Language server is starting up, and writes output. This flow goes from the UI Thread to a background thread, then switches back to the UI thread to write a log. The background thread can end up blocking waiting for the UI thread. This doesn't happen 100% of the time because we suspect it is a race condition around whether an awaited call returns to the UI thread or remains on a background thread.

We use `TaskFactory.Run` a lot to go from sync code to async code, and we often switch to the UI thread within these Run blocks. The takeaway is that we should not be calling TaskFactory.Run from code already running on a background thread. There are more instances of this throughout the codebase. Ideally we would transform everything to be async-first, but this will be a gradual process. This change focuses on the known reported issue.